Move SearchPage components into the SearchPage folder
