+++ src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Header from "./components/Header/Header.jsx";
 import Homepage from './components/Homepage/Homepage.jsx';
 import FavoritesPage from "./components/FavoritesPage/FavoritesPage.jsx";
-import SearchPage from "./components/SearchPage.jsx";
+import SearchPage from "./components/SearchPage/SearchPage.jsx";
 
 function App() {
   return (
